requirements.txt
aniso8601==8.0.0
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
Click==7.0
dominate==2.5.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pdfkit==0.6.1
psycopg2-binary==2.8.4
pycparser==2.20
pytz==2019.3
Safe==0.4
six==1.14.0
SQLAlchemy==1.3.13
visitor==0.1.3
Werkzeug==1.0.0
WTForms==2.2.1
pg8000==1.13.2; python_version > '3.0'
pg8000==1.12.5; python_version < '3.0'
gunicorn==20.0.4
Pillow==6.2.0
pymysql